+++ src/js/player.js
@@ -560,6 +560,7 @@ class Player extends Component {
     this.on(this.tech_, 'volumechange', this.handleTechVolumeChange_);
     this.on(this.tech_, 'texttrackchange', this.handleTechTextTrackChange_);
     this.on(this.tech_, 'loadedmetadata', this.updateStyleEl_);
+    this.on(this.tech_, 'posterchange', this.handleTechPosterChange_);
 
     this.usingNativeControls(this.techGet_('controls'));
 
@@ -671,6 +672,9 @@ class Player extends Component {
     if (this.cache_.volume) {
       this.techCall_('setVolume', this.cache_.volume);
     }
+
+    // Look if the tech found a higher resolution poster while loading
+    this.handleTechPosterChange_();
 
     // Update the duration if available
     this.handleTechDurationChange_();
@@ -1876,7 +1880,8 @@ class Player extends Component {
   }
 
   /**
-   * get or set the poster image source url
+   * Get or set the poster image source url
+   *
    * ##### EXAMPLE:
    * ```js
    *     // get
@@ -1911,6 +1916,26 @@ class Player extends Component {
     this.trigger('posterchange');
 
     return this;
+  }
+
+  /**
+   * Some techs (e.g. YouTube) can provide a poster source in an
+   * asynchronous way. We want the poster component to use this
+   * poster source so that it covers up the tech's controls.
+   * (YouTube's play button). However we only want to use this
+   * soruce if the player user hasn't set a poster through
+   * the normal APIs.
+   *
+   * @private
+   * @method handleTechPosterChange_
+   */
+  handleTechPosterChange_() {
+    if (!this.poster_ && this.tech_ && this.tech_.poster) {
+      this.poster_ = this.tech_.poster() || '';
+
+      // Let components know the poster has changed
+      this.trigger('posterchange');
+    }
   }
 
   /**