+++ CHANGELOG.md
@@ -146,6 +146,7 @@ CHANGELOG
 * @misteroneill updated play control to use its state for icon ([view](https://github.com/videojs/video.js/pull/2636))
 * @gkatsev exposed isCrossOrigin and used it to enable CORS for textTrack XHRs ([view](https://github.com/videojs/video.js/pull/2633))
 * @misteroneill fixed tsml to be used as a tag for template strings ([view](https://github.com/videojs/video.js/pull/2629))
+* @eXon added support for a tech-supplied poster ([view](https://github.com/videojs/video.js/pull/2339))
 
 --------------------